backend/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.135.1",
    "groq>=1.0.0",
    "httpx>=0.28.1",
    "pydantic-settings>=2.13.1",
]
